Moving template store to the object directory
