+++ tests/test_updater_multiple_fitters.py
@@ -5,7 +5,7 @@ import circusort
 import logging
 import scipy
 from circusort.io.utils import generate_fake_probe
-from circusort.io.template import TemplateStore
+from circusort.obj import TemplateStore
 
 
 host = '127.0.0.1' # to run the test locally
